raw_monitor/requirements.txt
paho-mqtt~=2.1
matplotlib~=3.10
numpy~=1.26
